+++ android/src/main/java/com/reactnativestripesdk/PaymentMethodCreateParamsFactory.kt
@@ -5,11 +5,12 @@ import com.stripe.android.model.*
 
 class PaymentMethodCreateParamsFactory(
   private val clientSecret: String,
-  private val params: ReadableMap,
+  private val paymentMethodData: ReadableMap?,
+  private val options: ReadableMap,
   private val cardFieldView: CardFieldView?,
   private val cardFormView: CardFormView?,
 ) {
-  private val billingDetailsParams = mapToBillingDetails(getMapOrNull(params, "billingDetails"), cardFieldView?.cardAddress ?: cardFormView?.cardAddress)
+  private val billingDetailsParams = mapToBillingDetails(getMapOrNull(paymentMethodData, "billingDetails"), cardFieldView?.cardAddress ?: cardFormView?.cardAddress)
 
   @Throws(PaymentMethodCreateParamsException::class)
   fun createConfirmParams(paymentMethodType: PaymentMethod.Type): ConfirmPaymentIntentParams {
@@ -62,18 +63,17 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createIDEALPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    val bankName = getValOr(params, "bankName", null)
+    val bankName = getValOr(paymentMethodData, "bankName", null)
 
     val idealParams = PaymentMethodCreateParams.Ideal(bankName)
     val createParams =
       PaymentMethodCreateParams.create(ideal = idealParams, billingDetails = billingDetailsParams)
-    val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage"))
 
     return ConfirmPaymentIntentParams
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = createParams,
         clientSecret = clientSecret,
-        setupFutureUsage = setupFutureUsage,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage")),
       )
   }
 
@@ -86,6 +86,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage")),
         )
     }
 
@@ -94,8 +95,8 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createCardPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    val paymentMethodId = getValOr(params, "paymentMethodId", null)
-    val token = getValOr(params, "token", null)
+    val paymentMethodId = getValOr(paymentMethodData, "paymentMethodId", null)
+    val token = getValOr(paymentMethodData, "token", null)
 
     val cardParams = cardFieldView?.cardParams ?: cardFormView?.cardParams
 
@@ -103,10 +104,10 @@ class PaymentMethodCreateParamsFactory(
       throw PaymentMethodCreateParamsException("Card details not complete")
     }
 
-    val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage"))
+    val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
 
     if (paymentMethodId != null) {
-      val cvc = getValOr(params, "cvc", null)
+      val cvc = getValOr(paymentMethodData, "cvc", null)
       val paymentMethodOptionParams =
         if (cvc != null) PaymentMethodOptionsParams.Card(cvc) else null
 
@@ -134,7 +135,7 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createIDEALPaymentSetupParams(): ConfirmSetupIntentParams {
-    val bankName = getValOr(params, "bankName", null)
+    val bankName = getValOr(paymentMethodData, "bankName", null)
 
     val idealParams = PaymentMethodCreateParams.Ideal(bankName)
     val createParams =
@@ -149,7 +150,7 @@ class PaymentMethodCreateParamsFactory(
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createSepaPaymentSetupParams(): ConfirmSetupIntentParams {
     billingDetailsParams?.let {
-      val iban = getValOr(params, "iban", null) ?: run {
+      val iban = getValOr(paymentMethodData, "iban", null) ?: run {
         throw PaymentMethodCreateParamsException("You must provide IBAN")
       }
 
@@ -188,10 +189,9 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createSofortPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    val country = getValOr(params, "country", null) ?: run {
+    val country = getValOr(paymentMethodData, "country", null) ?: run {
       throw PaymentMethodCreateParamsException("You must provide bank account country")
     }
-    val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage"))
 
     val params = PaymentMethodCreateParams.create(
       PaymentMethodCreateParams.Sofort(country = country),
@@ -202,13 +202,13 @@ class PaymentMethodCreateParamsFactory(
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = params,
         clientSecret = clientSecret,
-        setupFutureUsage = setupFutureUsage,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage")),
       )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createSofortPaymentSetupParams(): ConfirmSetupIntentParams {
-    val country = getValOr(params, "country", null)
+    val country = getValOr(paymentMethodData, "country", null)
       ?: throw PaymentMethodCreateParamsException("You must provide country")
 
     val params = PaymentMethodCreateParams.create(
@@ -231,20 +231,20 @@ class PaymentMethodCreateParamsFactory(
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = params,
         clientSecret = clientSecret,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage")),
       )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createBancontactPaymentConfirmParams(): ConfirmPaymentIntentParams {
     billingDetailsParams?.let {
-      val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage", null))
       val params = PaymentMethodCreateParams.createBancontact(it)
 
       return ConfirmPaymentIntentParams
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
-          setupFutureUsage = setupFutureUsage,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage")),
         )
     }
 
@@ -274,6 +274,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
         )
     }
 
@@ -289,6 +290,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
         )
     }
 
@@ -304,6 +306,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
         )
     }
 
@@ -313,10 +316,10 @@ class PaymentMethodCreateParamsFactory(
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createSepaPaymentConfirmParams(): ConfirmPaymentIntentParams {
     billingDetailsParams?.let {
-      val iban = getValOr(params, "iban", null) ?: run {
+      val iban = getValOr(paymentMethodData, "iban", null) ?: run {
         throw PaymentMethodCreateParamsException("You must provide IBAN")
       }
-      val setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage"))
+
       val params = PaymentMethodCreateParams.create(
         sepaDebit = PaymentMethodCreateParams.SepaDebit(iban),
         billingDetails = it
@@ -326,7 +329,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
-          setupFutureUsage = setupFutureUsage
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
         )
     }
 
@@ -335,7 +338,7 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createFpxPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    val bank = getBooleanOrFalse(params, "testOfflineBank").let { "test_offline_bank" }
+    val bank = getBooleanOrFalse(paymentMethodData, "testOfflineBank").let { "test_offline_bank" }
     val params = PaymentMethodCreateParams.create(
       PaymentMethodCreateParams.Fpx(bank)
     )
@@ -344,6 +347,7 @@ class PaymentMethodCreateParamsFactory(
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = params,
         clientSecret = clientSecret,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
       )
   }
 
@@ -356,6 +360,7 @@ class PaymentMethodCreateParamsFactory(
         .createWithPaymentMethodCreateParams(
           paymentMethodCreateParams = params,
           clientSecret = clientSecret,
+          setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
         )
     }
 
@@ -364,7 +369,7 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createAuBecsDebitPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    val formDetails = getMapOrNull(params, "formDetails") ?: run {
+    val formDetails = getMapOrNull(paymentMethodData, "formDetails") ?: run {
       throw PaymentMethodCreateParamsException("You must provide form details")
     }
 
@@ -390,12 +395,13 @@ class PaymentMethodCreateParamsFactory(
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = params,
         clientSecret = clientSecret,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
       )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createAuBecsDebitPaymentSetupParams(): ConfirmSetupIntentParams {
-    val formDetails = getMapOrNull(params, "formDetails") ?: run {
+    val formDetails = getMapOrNull(paymentMethodData, "formDetails") ?: run {
       throw PaymentMethodCreateParamsException("You must provide form details")
     }
 
@@ -426,16 +432,17 @@ class PaymentMethodCreateParamsFactory(
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createUSBankAccountPaymentSetupParams(): ConfirmSetupIntentParams {
-    params.getString("accountNumber")?.let {
+    // If payment method data is supplied, assume they are passing in the bank details manually
+    paymentMethodData?.let {
       if (billingDetailsParams?.name.isNullOrBlank()) {
         throw PaymentMethodCreateParamsException("When creating a US bank account payment method, you must provide the following billing details: name")
       }
       return ConfirmSetupIntentParams.create(
-        paymentMethodCreateParams = createUSBankAccountParams(params),
+        paymentMethodCreateParams = createUSBankAccountParams(paymentMethodData),
         clientSecret = clientSecret,
       )
     } ?: run {
-      // Payment method is assumed to be already attached through via collectBankAccountForSetup
+      // Payment method is assumed to be already attached through via collectBankAccount
       return ConfirmSetupIntentParams.create(
         clientSecret = clientSecret,
         paymentMethodType = PaymentMethod.Type.USBankAccount
@@ -459,23 +466,25 @@ class PaymentMethodCreateParamsFactory(
       .createWithPaymentMethodCreateParams(
         paymentMethodCreateParams = params,
         clientSecret = clientSecret,
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
       )
   }
 
   @Throws(PaymentMethodCreateParamsException::class)
   private fun createUSBankAccountPaymentConfirmParams(): ConfirmPaymentIntentParams {
-    params.getString("accountNumber")?.let {
+    // If payment method data is supplied, assume they are passing in the bank details manually
+    paymentMethodData?.let {
       if (billingDetailsParams?.name.isNullOrBlank()) {
         throw PaymentMethodCreateParamsException("When creating a US bank account payment method, you must provide the following billing details: name")
       }
 
       return ConfirmPaymentIntentParams.createWithPaymentMethodCreateParams(
-        paymentMethodCreateParams = createUSBankAccountParams(params),
+        paymentMethodCreateParams = createUSBankAccountParams(paymentMethodData),
         clientSecret,
-        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(params, "setupFutureUsage"))
+        setupFutureUsage = mapToPaymentIntentFutureUsage(getValOr(options, "setupFutureUsage"))
       )
     } ?: run {
-      // Payment method is assumed to be already attached through via collectBankAccountForPayment
+      // Payment method is assumed to be already attached through via collectBankAccount
       return ConfirmPaymentIntentParams.create(
         clientSecret = clientSecret,
         paymentMethodType = PaymentMethod.Type.USBankAccount