+++ ios/Mappers.swift
@@ -609,7 +609,7 @@ class Mappers {
         ]
         let method: NSDictionary = [
             "id": paymentMethod.stripeId,
-            "type": Mappers.mapPaymentMethodType(type: paymentMethod.type),
+            "paymentMethodType": Mappers.mapPaymentMethodType(type: paymentMethod.type),
             "livemode": paymentMethod.liveMode,
             "customerId": paymentMethod.customerId ?? NSNull(),
             "billingDetails": Mappers.mapFromBillingDetails(billingDetails: paymentMethod.billingDetails),